+++ ClientFactory.ts
@@ -101,4 +101,14 @@ export class ClientFactory {
       .withHttpMiddleware(httpMiddlewareOptions)
       .withClientCredentialsFlow(authMiddlewareOptions);
   }
+
+  static factorWithoutAuthenticationFlow: (clientConfig: ClientConfig) => Client = (clientConfig: ClientConfig) => {
+    const httpMiddlewareOptions: HttpMiddlewareOptions = {
+      host: clientConfig.authUrl,
+      headersWithStringBody: ['application/x-www-form-urlencoded'],
+      fetch,
+    };
+
+    return new ClientBuilder().withHttpMiddleware(httpMiddlewareOptions).build();
+  };
 }