+++ actionControllers/CartController.ts
@@ -19,11 +19,14 @@ import handleError from '@Commerce-commercetools/utils/handleError';
 import { SortAttributes, SortOrder } from '@Types/query/ProductQuery';
 import { OrderQuery } from '@Types/cart';
 import { fetchAccountFromSession } from '@Commerce-commercetools/utils/fetchAccountFromSession';
+import { CartNotMatchOrderError } from '@Commerce-commercetools/errors/CartNotMatchOrderError';
+import { TokenError } from '@Commerce-commercetools/errors/TokenError';
+import { Token } from '@Types/Token';
 
 type ActionHook = (request: Request, actionContext: ActionContext) => Promise<Response>;
 
 function getCartApi(request: Request, actionContext: ActionContext) {
-  return new CartApi(actionContext.frontasticContext, getLocale(request), getCurrency(request));
+  return new CartApi(actionContext.frontasticContext, getLocale(request), getCurrency(request), request);
 }
 
 function queryParamsToSortAttributes(queryParams: any) {
@@ -79,7 +82,8 @@ export const getCart: ActionHook = async (request: Request, actionContext: Actio
       statusCode: 200,
       body: cart ? JSON.stringify(cart) : '',
       sessionData: {
-        ...request.sessionData,
+        ...cartApi.getSessionData(),
+        ...(cart ? { cartId: cart.cartId } : {}),
       },
     };
   } catch (error) {
@@ -128,7 +132,7 @@ export const addToCart: ActionHook = async (request: Request, actionContext: Act
     statusCode: 200,
     body: JSON.stringify(cart),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId,
     },
   };
@@ -145,7 +149,7 @@ export const replicateCart: ActionHook = async (request: Request, actionContext:
       statusCode: 422,
       body: JSON.stringify(`Order was not found.`),
       sessionData: {
-        ...request.sessionData,
+        ...cartApi.getSessionData(),
       },
     };
   }
@@ -158,7 +162,7 @@ export const replicateCart: ActionHook = async (request: Request, actionContext:
         statusCode: 400,
         body: JSON.stringify(`We could not replicate cart for order : "${orderId}".`),
         sessionData: {
-          ...request.sessionData,
+          ...cartApi.getSessionData(),
         },
       };
     }
@@ -167,7 +171,7 @@ export const replicateCart: ActionHook = async (request: Request, actionContext:
       statusCode: 200,
       body: JSON.stringify(cart),
       sessionData: {
-        ...request.sessionData,
+        ...cartApi.getSessionData(),
         cartId: cart.cartId,
       },
     };
@@ -177,7 +181,7 @@ export const replicateCart: ActionHook = async (request: Request, actionContext:
         statusCode: error.status,
         body: JSON.stringify(error.message),
         sessionData: {
-          ...request.sessionData,
+          ...cartApi.getSessionData(),
         },
       };
     }
@@ -186,7 +190,7 @@ export const replicateCart: ActionHook = async (request: Request, actionContext:
       statusCode: 400,
       body: JSON.stringify(err.message),
       sessionData: {
-        ...request.sessionData,
+        ...cartApi.getSessionData(),
       },
     };
   }
@@ -213,7 +217,7 @@ export const updateLineItem: ActionHook = async (request: Request, actionContext
     statusCode: 200,
     body: JSON.stringify(cart),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId,
     },
   };
@@ -241,7 +245,7 @@ export const removeLineItem: ActionHook = async (request: Request, actionContext
     statusCode: 200,
     body: JSON.stringify(cart),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId,
     },
   };
@@ -259,7 +263,7 @@ export const updateCart: ActionHook = async (request: Request, actionContext: Ac
     statusCode: 200,
     body: JSON.stringify(cart),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId,
     },
   };
@@ -286,7 +290,7 @@ export const checkout: ActionHook = async (request: Request, actionContext: Acti
     statusCode: 200,
     body: JSON.stringify(order),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId,
     },
   };
@@ -312,7 +316,7 @@ export const getOrders: ActionHook = async (request: Request, actionContext: Act
     statusCode: 200,
     body: JSON.stringify(orders),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
     },
   };
   return response;
@@ -328,7 +332,7 @@ export const getShippingMethods: ActionHook = async (request: Request, actionCon
     statusCode: 200,
     body: JSON.stringify(shippingMethods),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
     },
   };
 
@@ -345,7 +349,7 @@ export const getAvailableShippingMethods: ActionHook = async (request: Request, 
     statusCode: 200,
     body: JSON.stringify(availableShippingMethods),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId: cart.cartId,
     },
   };
@@ -371,7 +375,7 @@ export const setShippingMethod: ActionHook = async (request: Request, actionCont
     statusCode: 200,
     body: JSON.stringify(cart),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId: cart.cartId,
     },
   };
@@ -407,7 +411,7 @@ export const addPaymentByInvoice: ActionHook = async (request: Request, actionCo
     statusCode: 200,
     body: JSON.stringify(cart),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId: cart.cartId,
     },
   };
@@ -429,7 +433,7 @@ export const updatePayment: ActionHook = async (request: Request, actionContext:
     statusCode: 200,
     body: JSON.stringify(payment),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId: cart.cartId,
     },
   };
@@ -454,7 +458,7 @@ export const redeemDiscount: ActionHook = async (request: Request, actionContext
       statusCode: 200,
       body: JSON.stringify(cart),
       sessionData: {
-        ...request.sessionData,
+        ...cartApi.getSessionData(),
         cartId: cart.cartId,
       },
     };
@@ -464,7 +468,7 @@ export const redeemDiscount: ActionHook = async (request: Request, actionContext
         statusCode: error.status,
         body: JSON.stringify(error.message),
         sessionData: {
-          ...request.sessionData,
+          ...cartApi.getSessionData(),
           cartId: cart.cartId,
         },
       };
@@ -496,7 +500,7 @@ export const removeDiscount: ActionHook = async (request: Request, actionContext
     statusCode: 200,
     body: JSON.stringify(cart),
     sessionData: {
-      ...request.sessionData,
+      ...cartApi.getSessionData(),
       cartId: cart.cartId,
     },
   };
@@ -530,11 +534,100 @@ export const queryOrders: ActionHook = async (request, actionContext) => {
     const response: Response = {
       statusCode: 200,
       body: JSON.stringify(queryResult),
-      sessionData: request.sessionData,
+      sessionData: {
+        ...cartApi.getSessionData(),
+      },
     };
 
     return response;
   } catch (error) {
     return handleError(error, request);
   }
+};
+
+export const getCheckoutOrder: ActionHook = async (request, actionContext) => {
+  const locale = getLocale(request);
+  const cartApi = new CartApi(actionContext.frontasticContext, locale, getCurrency(request));
+
+  const account = fetchAccountFromSession(request);
+
+  try {
+    const orderQuery: OrderQuery = {
+      accountId: account?.accountId,
+      orderIds: [request.query?.orderId],
+      limit: 1,
+    };
+
+    const queryResult = await cartApi.queryOrders(orderQuery);
+
+    // We'll consider the first order as the checkout order
+    const order = queryResult.items[0];
+
+    if (account === undefined) {
+      // If account is not logged in, we need to validate if the order belongs to the current session cart
+      if (order?.cartId !== request.sessionData?.cartId) {
+        throw new CartNotMatchOrderError({ message: 'Order does not match the current cart.' });
+      }
+    }
+
+    const response: Response = {
+      statusCode: 200,
+      body: JSON.stringify(order),
+      sessionData: {
+        ...cartApi.getSessionData(),
+      },
+    };
+
+    return response;
+  } catch (error) {
+    return handleError(error, request);
+  }
+};
+
+export const getCheckoutToken: ActionHook = async (request: Request, actionContext: ActionContext) => {
+  let checkoutToken: Token;
+
+  const cartApi = getCartApi(request, actionContext);
+  const cart = await CartFetcher.fetchCart(cartApi, request, actionContext);
+  const account = fetchAccountFromSession(request);
+
+  try {
+    checkoutToken = await cartApi.getCheckoutToken(cart, account);
+  } catch (error) {
+    if (error instanceof TokenError) {
+      const response: Response = {
+        statusCode: 401,
+        body: JSON.stringify(error.message),
+        sessionData: {
+          ...cartApi.getSessionData(),
+        },
+      };
+
+      return response;
+    }
+
+    if (error instanceof ExternalError) {
+      const response: Response = {
+        statusCode: error.status,
+        body: JSON.stringify(error.message),
+        sessionData: {
+          ...cartApi.getSessionData(),
+        },
+      };
+
+      return response;
+    }
+
+    throw error;
+  }
+
+  const response: Response = {
+    statusCode: 200,
+    body: checkoutToken ? JSON.stringify(checkoutToken) : '',
+    sessionData: {
+      ...cartApi.getSessionData(),
+    },
+  };
+
+  return response;
 };