+++ apis/CartApi.ts
@@ -37,19 +37,24 @@ import { ExternalError } from '../utils/Errors';
 import { CartNotCompleteError } from '../errors/CartNotCompleteError';
 import { CartPaymentNotFoundError } from '../errors/CartPaymentNotFoundError';
 import { CartRedeemDiscountCodeError } from '../errors/CartRedeemDiscountCodeError';
-import { Context } from '@frontastic/extension-types';
+import { Context, Request } from '@frontastic/extension-types';
 import { ProductApi } from './ProductApi';
 import { ProductMapper } from '@Commerce-commercetools/mappers/ProductMapper';
 import { getOffsetFromCursor } from '@Commerce-commercetools/utils/Pagination';
 import { PaginatedResult } from '@Types/result';
 import { OrderQuery } from '@Types/cart';
+import { CartNotActiveError } from '@Commerce-commercetools/errors/CartNotActiveError';
+import { Token } from '@Types/Token';
+import { tokenHasExpired } from '@Commerce-commercetools/utils/Token';
+import { CartSetAnonymousIdAction } from '@commercetools/platform-sdk/dist/declarations/src/generated/models/cart';
+import { TokenError } from '@Commerce-commercetools/errors/TokenError';
 
 export class CartApi extends BaseApi {
   productApi: ProductApi;
 
-  constructor(frontasticContext: Context, locale: string | null, currency: string | null) {
-    super(frontasticContext, locale, currency);
-    this.productApi = new ProductApi(frontasticContext, locale, currency);
+  constructor(frontasticContext: Context, locale: string | null, currency: string | null, request?: Request | null) {
+    super(frontasticContext, locale, currency, request);
+    this.productApi = new ProductApi(frontasticContext, locale, currency, request);
   }
 
   replicateCart: (orderId: string) => Promise<Cart> = async (orderId: string) => {
@@ -87,6 +92,7 @@ export class CartApi extends BaseApi {
             'paymentInfo.payments[*]',
           ],
           where: [`customerId="${account.accountId}"`, `cartState="Active"`],
+          sort: 'lastModifiedAt desc',
         },
       })
       .execute()
@@ -127,7 +133,7 @@ export class CartApi extends BaseApi {
       });
   };
 
-  getAnonymous: (anonymousId: string) => Promise<Cart> = async (anonymousId: string) => {
+  getAnonymous: () => Promise<Cart> = async () => {
     const locale = await this.getCommercetoolsLocal();
 
     const response = await this.requestBuilder()
@@ -140,7 +146,7 @@ export class CartApi extends BaseApi {
             'discountCodes[*].discountCode',
             'paymentInfo.payments[*]',
           ],
-          where: [`anonymousId="${anonymousId}"`, `cartState="Active"`],
+          where: [`anonymousId="${this.getAnonymousIdFromSessionData()}"`, `cartState="Active"`],
           sort: 'createdAt desc',
         },
       })
@@ -152,6 +158,13 @@ export class CartApi extends BaseApi {
     if (response.body.count >= 1) {
       return this.buildCartWithAvailableShippingMethods(response.body.results[0], locale);
     }
+
+    // If there is no active cart, we create one with new anonymousId and checkout token
+    this.invalidateSessionAnonymousId();
+    const anonymousId = this.getAnonymousIdFromSessionData();
+
+    // Before create new cart with the anonymousId, we need to get a checkout token for the anonymousId
+    await this.generateCheckoutToken(anonymousId);
 
     const cartDraft: CartDraft = {
       currency: locale.currency,
@@ -182,7 +195,7 @@ export class CartApi extends BaseApi {
       });
   };
 
-  getById: (cartId: string) => Promise<Cart> = async (cartId: string) => {
+  getActiveCartById: (cartId: string) => Promise<Cart> = async (cartId: string) => {
     const locale = await this.getCommercetoolsLocal();
 
     return await this.requestBuilder()
@@ -202,6 +215,9 @@ export class CartApi extends BaseApi {
       })
       .execute()
       .then((response) => {
+        if (response.body.cartState !== 'Active') {
+          throw new CartNotActiveError({ message: `Cart ${cartId} is not active.` });
+        }
         return this.buildCartWithAvailableShippingMethods(response.body, locale);
       })
       .catch((error) => {
@@ -958,7 +974,11 @@ export class CartApi extends BaseApi {
       sortAttributes.push(`lastModifiedAt desc`);
     }
 
-    const whereClause = [`customerId="${orderQuery.accountId}"`];
+    const whereClause = [];
+
+    if (orderQuery.accountId !== undefined) {
+      whereClause.push(`customerId="${orderQuery.accountId}"`);
+    }
 
     if (orderQuery.orderIds !== undefined && orderQuery.orderIds.length !== 0) {
       whereClause.push(`id in ("${orderQuery.orderIds.join('","')}")`);
@@ -1008,5 +1028,55 @@ export class CartApi extends BaseApi {
       .catch((error) => {
         throw new ExternalError({ status: error.code, message: error.message, body: error.body });
       });
+  }
+
+  async getCheckoutToken(cart: Cart, account?: Account): Promise<Token | undefined> {
+    let checkoutToken = this.getSessionCheckoutToken();
+
+    if (!tokenHasExpired(checkoutToken)) {
+      return checkoutToken;
+    }
+
+    if (checkoutToken.refreshToken) {
+      await this.generateCheckoutToken(undefined, undefined, checkoutToken.refreshToken).catch((error) => {
+        throw new ExternalError({ status: error.code, message: error.message, body: error.body });
+      });
+
+      return this.getSessionCheckoutToken();
+    }
+
+    // The token has expired as we can't refresh it, we need to create a new one
+    this.invalidateSessionCheckoutData();
+
+    if (account) {
+      // If the user is logged in, we can't create a new token without the user email and password
+      throw new TokenError({
+        message: 'The checkout token has expired and can not be refreshed. Please login again.',
+      });
+    }
+
+    const anonymousId = this.getAnonymousIdFromSessionData();
+
+    await this.generateCheckoutToken(anonymousId).catch((error) => {
+      throw new ExternalError({ status: error.code, message: error.message, body: error.body });
+    });
+
+    checkoutToken = this.getSessionCheckoutToken();
+
+    // Update the cart with the new anonymousId
+    const cartUpdate: CartUpdate = {
+      version: +cart.cartVersion,
+      actions: [
+        {
+          action: 'setAnonymousId',
+          anonymousId,
+        } as CartSetAnonymousIdAction,
+      ],
+    };
+
+    const locale = await this.getCommercetoolsLocal();
+    await this.updateCart(cart.cartId, cartUpdate, locale);
+
+    return checkoutToken;
   }
 }